XL: Add pid for revised streamdeck xl
We encountered a streamdeck xl with a new pid 0x008f
which will work fine if treated like a normal xl.
This commit adds the pid to the .rules and a
const including mapping to Kind::XL in lib.rs
Since we don't have access to the device, there is
no further information about it.